Create README.md
This is the Visual Basic implementation of Face Recognition [from a picture, not a webcam/video] using EmguCV. I have seen a lot of examples for C# , and a video , but couldn't find a simple sample for a still picture, and Visual Basic  2019, so I have tried to make my own. This is partly (or mostly ? 🥇 ) [inspired from https://github.com/mesutpiskin] 's work. Performance was not my priority , and I believe should be worked on. Any input would be appreciated on that (also others)regard.